requirement.txt
fastapi
uvicorn
langchain
langchain-text-splitters
langchain-community
langchain-google-genai
langchain-chroma
pypdf
openai
python-dotenv
